임시저장

+++ easyBack/src/main/java/com/ssafy/easyback/neighbor/model/service/NeighborServiceImpl.java
@@ -25,6 +25,7 @@ public class NeighborServiceImpl implements NeighborService{
 
     @Override
     public void deleteNeighbor(NeighborDto neighborDto) throws Exception {
-        neighborMapper.deleteNeighbor(neighborDto);
+        if(neighborDto.getStatus().equals("NEIGHBOR"))  neighborMapper.deleteNeighbor(neighborDto);
+        else neighborMapper.cancelRequest(neighborDto);
     }
 }
